Fixing event and registration relation
