+++ app/Models/Event.php
@@ -18,30 +18,6 @@ class Event extends Model
 
     // public $timestamps = false;
 
-    public function scopeVisibleToUser(Builder $query, $user, array $filters)
-    {
-        $query->where(fn ($query) =>
-            $query->whereIn('event_level', ['university', 'international', 'regional', 'national'])
-
-                ->orWhere(fn ($query) =>
-                    $query->where('event_level', 'faculty')//Cek apabila event levelnya faculty/facultas
-                        ->where('faculty_id', $user->faculty_id)
-                )
-
-                ->orWhere(fn ($query) =>
-                    $query->where('event_level', 'major') //Cek apabila event levelnya major/prodi
-                        ->where('major_id', $user->major_id)
-                )
-        );
-
-        $query->when($filters['search']??false, fn ($query, $filters) => $query->where('name', 'like', '%'. request('search').'%'));
-
-
-        return $query;
-
-
-    }
-
 
     public function users(): BelongsToMany
     {
@@ -101,5 +77,10 @@ class Event extends Model
     public function contacts(): HasMany
     {
         return $this->hasMany(ContactPerson::class);
+    }
+
+    public function registration(): BelongsTo
+    {
+        return $this->belongsTo(EventRegistration::class);
     }
 }
